Fix Vagrant environment and documentation
Fix inconsistencies in Vagrant config template and installation scripts
introduced by the move to the Git-based content repository.

Fix systemd service description to match user names, bug introduced by #860

Add dummy paypal and credit card configuration to config template to
make it more obvious what needs to be configured.

Update README to reflect the changes, make dependcies more obvious and
document the changes introduced by #862

Avoid overriding existing configuration files, if one exists.

+++ build/vagrant/configure_app.sh
@@ -8,11 +8,9 @@ systemctl daemon-reload
 systemctl start fundraising_app.service
 
 # Configure app
-if [[ -z "$WIKI_PASSWD" ]]; then
-    WIKI_PASSWD="WIKI PASSWORD MISSING!!!"
+if [ ! -f /vagrant/app/config/config.prod.json ]; then
+    cp /vagrant/build/vagrant/config.prod.json /vagrant/app/config/config.prod.json
 fi
-cp /vagrant/build/vagrant/config.prod.json /vagrant/app/config/config.prod.json
-sed -i -e "s/__WIKI_PASSWORD__/$WIKI_PASSWD/" /vagrant/app/config/config.prod.json
 
 # Log all outgoing mails instead of sending them
 echo "cat >> /tmp/logmail.log" > /usr/local/bin/logmail
